update numpy 1.24 (#7552)

+++ deploy/python/utils.py
@@ -357,7 +357,7 @@ def nms(dets, match_threshold=0.6, match_metric='iou'):
     order = scores.argsort()[::-1]
 
     ndets = dets.shape[0]
-    suppressed = np.zeros((ndets), dtype=np.int)
+    suppressed = np.zeros((ndets), dtype=np.int32)
 
     for _i in range(ndets):
         i = order[_i]
